configure LobbiesSocketsController to understand websockets-rails
As a first step to figuring out how to use the websockets-rails gem this
is just an exploration of the bare-minimum needed to use their
javascript dispatcher. LobbiesSocketsController should likely be name
after the players it connects, and there was a lot of configuring on the
gem front in order to get everything working.